Disable completion for krew plugin builds
